chore: clean up handlers and orchestrator organization
handlers/ cleanup:
- Rename handler_*_test.go → *_test.go (removes redundant prefix)
- Merge duplicate test files (stealth, upload)
- Remove handlers_extra_test.go (redistribute to proper files)
- Split navigation.go → navigation.go + evaluate.go
- Split media.go → screenshot.go + pdf.go + text.go

orchestrator/ refactor:
- Replace 15 boilerplate handleTab* functions with generic proxyTabRequest
- Split handlers.go (855 lines) into:
  - handlers.go (62 lines) - registry only
  - proxy.go (181 lines) - generic proxy functions
  - handlers_instances.go (182 lines) - instance CRUD
  - handlers_profiles.go (93 lines) - profile management
  - handlers_tabs.go (108 lines) - custom tab handlers
- Reduced ~230 lines of duplicate code
