--- internal/handlers/handler_upload_test.go
@@ -1,81 +1,0 @@
-package handlers
-
-import (
-	"bytes"
-	"net/http"
-	"net/http/httptest"
-	"testing"
-
-	"github.com/pinchtab/pinchtab/internal/config"
-)
-
-func TestHandleUpload_BadJSON(t *testing.T) {
-	h := New(&mockBridge{}, &config.RuntimeConfig{}, nil, nil, nil)
-	req := httptest.NewRequest("POST", "/upload", bytes.NewReader([]byte("not json")))
-	req.Header.Set("Content-Type", "application/json")
-	w := httptest.NewRecorder()
-	h.HandleUpload(w, req)
-	if w.Code != http.StatusBadRequest {
-		t.Errorf("expected 400 for bad JSON, got %d", w.Code)
-	}
-}
-
-func TestHandleUpload_EmptyPaths(t *testing.T) {
-	h := New(&mockBridge{}, &config.RuntimeConfig{}, nil, nil, nil)
-	body := `{"selector": "input[type=file]"}`
-	req := httptest.NewRequest("POST", "/upload", bytes.NewReader([]byte(body)))
-	req.Header.Set("Content-Type", "application/json")
-	w := httptest.NewRecorder()
-	h.HandleUpload(w, req)
-	if w.Code != http.StatusBadRequest {
-		t.Errorf("expected 400 for empty paths, got %d", w.Code)
-	}
-}
-
-func TestHandleUpload_NonexistentPath(t *testing.T) {
-	h := New(&mockBridge{}, &config.RuntimeConfig{}, nil, nil, nil)
-	body := `{"selector": "input[type=file]", "paths": ["/tmp/nonexistent-file-12345.jpg"]}`
-	req := httptest.NewRequest("POST", "/upload", bytes.NewReader([]byte(body)))
-	req.Header.Set("Content-Type", "application/json")
-	w := httptest.NewRecorder()
-	h.HandleUpload(w, req)
-	if w.Code != http.StatusBadRequest {
-		t.Errorf("expected 400 for nonexistent path, got %d", w.Code)
-	}
-}
-
-func TestHandleTabUpload_MissingTabID(t *testing.T) {
-	h := New(&mockBridge{}, &config.RuntimeConfig{}, nil, nil, nil)
-	req := httptest.NewRequest("POST", "/tabs//upload", bytes.NewReader([]byte(`{"selector":"input[type=file]"}`)))
-	req.Header.Set("Content-Type", "application/json")
-	w := httptest.NewRecorder()
-	h.HandleTabUpload(w, req)
-	if w.Code != http.StatusBadRequest {
-		t.Errorf("expected 400, got %d", w.Code)
-	}
-}
-
-func TestHandleTabUpload_NoTab(t *testing.T) {
-	h := New(&mockBridge{failTab: true}, &config.RuntimeConfig{}, nil, nil, nil)
-	req := httptest.NewRequest("POST", "/tabs/tab_abc/upload", bytes.NewReader([]byte(`{"files":["aGVsbG8="]}`)))
-	req.SetPathValue("id", "tab_abc")
-	req.Header.Set("Content-Type", "application/json")
-	w := httptest.NewRecorder()
-	h.HandleTabUpload(w, req)
-	if w.Code != http.StatusNotFound {
-		t.Errorf("expected 404, got %d", w.Code)
-	}
-}
-
-func TestHandleUpload_BodyTooLarge(t *testing.T) {
-	h := New(&mockBridge{}, &config.RuntimeConfig{}, nil, nil, nil)
-	// Create a body larger than 10MB
-	bigBody := make([]byte, 11<<20) // 11MB
-	req := httptest.NewRequest("POST", "/upload", bytes.NewReader(bigBody))
-	req.Header.Set("Content-Type", "application/json")
-	w := httptest.NewRecorder()
-	h.HandleUpload(w, req)
-	if w.Code != http.StatusBadRequest {
-		t.Errorf("expected 400 for oversized body, got %d", w.Code)
-	}
-}
