+++ internal/handlers/snapshot_test.go
